requirements_selenium.txt
selenium>=4.15.0
pyperclip>=1.8.2
webdriver-manager>=4.0.1
Pillow>=10.0.0
